third_party/WebKit/LayoutTests/media/video-currentTime-before-have-metadata.html

<!DOCTYPE html>
<title>Test currentTime values when setting while readystate is HAVE_NOTHING.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-file.js"></script>
<video></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    video.src = findMediaFile("video", "content/test");
    assert_equals(video.currentTime, 0);
    video.currentTime = 1;
    assert_equals(video.currentTime, 1);

    video.onloadedmetadata = t.step_func(function() {
        assert_equals(video.currentTime, 1);
    });

    video.onseeked = t.step_func_done();
});
</script>
